Move RecordingExceptionHandler to flink/test-util
